@@ -8,7 +8,7 @@ use egui::{Frame, Margin, RichText};
 use re_dataframe_ui::{ColumnBlueprint, default_display_name_for_column};
 use re_log_types::{EntityPathPart, EntryId};
 use re_protos::cloud::v1alpha1::{EntryKind, ScanSegmentTableResponse};
-use re_redap_client::ConnectionRegistryHandle;
+use re_redap_client::{ConnectionRegistryHandle, Credentials};
 use re_sorbet::ColumnDescriptorRef;
 use re_ui::alert::Alert;
 use re_ui::{UiExt as _, icons};
@@ -394,6 +394,20 @@ impl RedapServers {
 
     pub fn logout(&mut self) {
         self.server_modal_ui.logout();
+        // Log out from the servers that used the accounts token.
+        for server in self.servers.values() {
+            if matches!(
+                server.connection_registry.credentials(&server.origin),
+                Some(Credentials::Stored)
+            ) {
+                server
+                    .connection_registry
+                    .remove_credentials(&server.origin);
+                self.command_sender
+                    .send(Command::RefreshCollection(server.origin.clone()))
+                    .ok();
+            }
+        }
     }
 
     /// Per-frame housekeeping.
@@ -561,7 +575,7 @@ impl RedapServers {
         let result = self.command_sender.send(command);
 
         if let Err(err) = result {
-            re_log::warn_once!("Failed to send command: {}", err);
+            re_log::warn_once!("Failed to send command: {err}");
         }
     }
 
