@@ -1,24 +1,22 @@
 use std::collections::BTreeSet;
-use std::collections::btree_map::Entry as BTreeMapEntry;
-use std::collections::hash_map::Entry as HashMapEntry;
+use std::sync::Arc;
 use std::time::Duration;
 
-use ahash::{HashMap, HashSet};
 use nohash_hasher::IntMap;
 use web_time::Instant;
 
 use re_byte_size::SizeBytes;
-use re_chunk::{Chunk, ChunkId, ComponentIdentifier, TimelineName};
-use re_log_types::{AbsoluteTimeRange, EntityPath, TimeInt};
+use re_chunk::{Chunk, ChunkId, TimelineName};
+use re_log_types::{AbsoluteTimeRange, TimeInt};
 
-// Used all over in docstrings.
-#[expect(unused_imports)]
-use crate::RowId;
-use crate::store::ChunkIdSetPerTime;
 use crate::{
     ChunkStore, ChunkStoreChunkStats, ChunkStoreDiff, ChunkStoreDiffKind, ChunkStoreEvent,
     ChunkStoreStats,
 };
+
+// Used all over in docstrings.
+#[expect(unused_imports)]
+use crate::RowId;
 
 // ---
 
@@ -29,7 +27,7 @@ pub enum GarbageCollectionTarget {
     /// The fraction must be a float in the range [0.0 : 1.0].
     DropAtLeastFraction(f64),
 
-    /// GC Everything that isn't protected.
+    /// GC Everything that isn't [protected](GarbageCollectionOptions::protect_latest).
     Everything,
 }
 
@@ -50,12 +48,24 @@ pub struct GarbageCollectionOptions {
     pub time_budget: Duration,
 
     /// How many component revisions to preserve on each timeline.
+    ///
+    /// This is ignored when using [`GarbageCollectionOptions::furthest_from`], unless the GC falls
+    /// back to row ID based collection.
     pub protect_latest: usize,
 
     /// Do not remove any data within these time ranges.
+    ///
+    /// This is ignored when using [`GarbageCollectionOptions::furthest_from`], unless the GC falls
+    /// back to row ID based collection.
     pub protected_time_ranges: IntMap<TimelineName, AbsoluteTimeRange>,
 
     /// Remove chunks giving priority to those that are the furthest away from this timestamp.
+    ///
+    /// This ignores [`protect_latest`] as well as [`protected_time_ranges`], unless the GC falls
+    /// back to row ID based collection.
+    ///
+    /// [`protect_latest`]: `GarbageCollectionOptions::protect_latest`
+    /// [`protected_time_ranges`]: `GarbageCollectionOptions::protected_time_ranges`
     pub furthest_from: Option<(TimelineName, TimeInt)>,
 }
 
@@ -94,19 +104,19 @@ impl std::fmt::Display for GarbageCollectionTarget {
     }
 }
 
-pub type RemovableChunkIdPerTimePerComponentPerTimelinePerEntity = IntMap<
-    EntityPath,
-    IntMap<TimelineName, IntMap<ComponentIdentifier, HashMap<TimeInt, Vec<ChunkId>>>>,
->;
-
 impl ChunkStore {
     /// Triggers a garbage collection according to the desired `target`.
     ///
-    /// Returns the list of `Chunk`s that were purged from the store in the form of [`ChunkStoreEvent`]s.
+    /// Returns the list of physical `Chunk`s that were purged from the store in the form
+    /// of [`ChunkStoreEvent`]s.
     ///
     /// ## Semantics
     ///
-    /// Garbage collection works on a chunk-level basis and is driven by [`RowId`] order
+    /// Garbage collection works on a chunk-level basis, giving priority to those that are the
+    /// furthest away from the timestamp specified in [`GarbageCollectionOptions::furthest_from`].
+    ///
+    /// If no timestamp is specified, or if not enough data could be collected during the
+    /// timestamp-driven pass, then garbage collection falls back to [`RowId`] order
     /// (specifically, the smallest `RowId` of each respective Chunk), i.e. the order defined
     /// by the clients' wall-clocks, allowing it to drop data across the different timelines in
     /// a fair, deterministic manner.
@@ -137,8 +147,6 @@ impl ChunkStore {
         let total_num_chunks_before = stats_before.total().num_chunks;
         let total_num_rows_before = stats_before.total().num_rows;
 
-        let protected_chunk_ids = self.find_all_protected_chunk_ids(options.protect_latest);
-
         let diffs = match options.target {
             GarbageCollectionTarget::DropAtLeastFraction(p) => {
                 assert!((0.0..=1.0).contains(&p));
@@ -158,8 +166,9 @@ impl ChunkStore {
                     "starting GC"
                 );
 
-                self.gc_drop_at_least_num_bytes(options, num_bytes_to_drop, &protected_chunk_ids)
+                self.gc_drop_at_least_num_bytes(options, num_bytes_to_drop)
             }
+
             GarbageCollectionTarget::Everything => {
                 re_log::trace!(
                     kind = "gc",
@@ -170,7 +179,7 @@ impl ChunkStore {
                     "starting GC"
                 );
 
-                self.gc_drop_at_least_num_bytes(options, f64::INFINITY, &protected_chunk_ids)
+                self.gc_drop_at_least_num_bytes(options, f64::INFINITY)
             }
         };
 
@@ -223,11 +232,14 @@ impl ChunkStore {
         (events, stats_before - stats_after)
     }
 
-    /// For each `EntityPath`, `Timeline`, `Component` find the N latest [`ChunkId`]s.
+    /// For each `EntityPath`, `Timeline`, `Component` find the N latest *physical* [`ChunkId`]s.
+    ///
+    /// This only accounts for physical/loaded chunks, and therefore will work properly even in the
+    /// context of a store that has offloaded some chunks at the end of its range.
     //
     // TODO(jleibs): More complex functionality might required expanding this to also
     // *ignore* specific entities, components, timelines, etc. for this protection.
-    fn find_all_protected_chunk_ids(&self, target_count: usize) -> BTreeSet<ChunkId> {
+    fn find_all_protected_physical_chunk_ids(&self, target_count: usize) -> BTreeSet<ChunkId> {
         re_tracing::profile_function!();
 
         if target_count == 0 {
@@ -243,21 +255,29 @@ impl ChunkStore {
                             |(_, temporal_chunk_ids_per_time)| {
                                 temporal_chunk_ids_per_time
                                     .per_start_time
-                                    .last_key_value()
-                                    .map(|(_, chunk_ids)| chunk_ids.iter().copied())
-                                    .into_iter()
+                                    .values()
+                                    .rev()
                                     .flatten()
+                                    .copied()
                                     .chain(
                                         temporal_chunk_ids_per_time
                                             .per_end_time
-                                            .last_key_value()
-                                            .map(|(_, chunk_ids)| chunk_ids.iter().copied())
-                                            .into_iter()
-                                            .flatten(),
+                                            .values()
+                                            .rev()
+                                            .flatten()
+                                            .copied(),
                                     )
-                                    .collect::<BTreeSet<_>>()
-                                    .into_iter()
-                                    .rev()
+                                    .filter(|chunk_id| {
+                                        self.chunks_per_chunk_id.contains_key(chunk_id) // make sure it's physical
+                                    })
+                                    // We might get unlucky and not hit the target count because `per_end_time`
+                                    // ended up yielding the same chunks as `per_start_time`, which will get
+                                    // deduplicated afterwards.
+                                    // This is fine for now, for two reasons:
+                                    // 1. In practice, we never use anything other than a `target_count` of 1, which
+                                    //    makes this whole thing irrelevant.
+                                    // 2. The whole concept of "protecting latest" only makes sense in the context
+                                    //    of the legacy data paths, which are on their way out anyhow.
                                     .take(target_count)
                             },
                         )
@@ -271,19 +291,14 @@ impl ChunkStore {
         &mut self,
         options: &GarbageCollectionOptions,
         mut num_bytes_to_drop: f64,
-        protected_chunk_ids: &BTreeSet<ChunkId>,
     ) -> Vec<ChunkStoreDiff> {
         re_tracing::profile_function!(re_format::format_bytes(num_bytes_to_drop));
 
-        let mut chunk_ids_dangling = HashSet::default();
-        let mut chunk_ids_to_be_removed =
-            RemovableChunkIdPerTimePerComponentPerTimelinePerEntity::default();
-
         let start_time = Instant::now();
-        {
+        let chunks_to_be_removed = {
             re_tracing::profile_scope!("mark");
 
-            // These chunks cannot be dangling by definition, since we need to access their data in
+            // These are all physical/loaded chunks by definition, since we need to access their data in
             // order to sort them in the first place.
             //
             // TODO(cmc): we would very much like that to be iterative or at least paginated in
@@ -307,80 +322,74 @@ impl ChunkStore {
                 vec![]
             };
 
-            let chunks_in_min_row_id_order =
+            let chunks_in_min_row_id_order = || {
+                // NOTE: latest-at protection only applies for RowID-based collection
+                let protected_chunk_ids =
+                    self.find_all_protected_physical_chunk_ids(options.protect_latest);
+
                 self.chunk_ids_per_min_row_id
-                    .iter()
-                    .filter_map(|(_, chunk_id)| {
-                        if let Some(chunk) = self.chunks_per_chunk_id.get(chunk_id) {
-                            Some(chunk.clone())
-                        } else {
-                            chunk_ids_dangling.insert(*chunk_id);
-                            None
-                        }
-                    });
+                    .values()
+                    .filter(move |chunk_id| !protected_chunk_ids.contains(chunk_id))
+                    .filter_map(|chunk_id| self.chunks_per_chunk_id.get(chunk_id).cloned()) // physical only
+                    .filter(|chunk| !chunk.is_static()) // cannot gc static data
+            };
 
             let chunks_in_priority_order = chunks_furthest_away
                 .into_iter()
-                .chain(chunks_in_min_row_id_order);
+                // `find_all_protected_physical_chunk_ids` is rather expensive so make sure we only
+                // compute it if we couldn't find enough chunks to remove already.
+                // Do note that `chunks_in_min_row_id_order` is a callback!
+                .chain(std::iter::once_with(chunks_in_min_row_id_order).flatten())
+                .filter(|chunk| !options.is_chunk_temporally_protected(chunk));
 
-            for chunk in
-                chunks_in_priority_order.filter(|chunk| !protected_chunk_ids.contains(&chunk.id()))
-            {
-                if options.is_chunk_temporally_protected(&chunk) {
-                    continue;
-                }
-
+            let mut chunks_to_be_removed = Vec::new();
+            for chunk in chunks_in_priority_order {
                 // NOTE: Do _NOT_ use `chunk.total_size_bytes` as it is sitting behind an Arc
                 // and would count as amortized (i.e. 0 bytes).
                 num_bytes_to_drop -= <Chunk as SizeBytes>::total_size_bytes(&*chunk) as f64;
 
-                // NOTE: We cannot blindly `retain` across all temporal tables, it's way too costly
-                // and slow. Rather we need to surgically remove the superfluous chunks.
-                let entity_path = chunk.entity_path();
-                let per_timeline = chunk_ids_to_be_removed
-                    .entry(entity_path.clone())
-                    .or_default();
-                for (&timeline, time_column) in chunk.timelines() {
-                    let per_component = per_timeline.entry(timeline).or_default();
-                    for component in chunk.components_identifiers() {
-                        let per_time = per_component.entry(component).or_default();
+                chunks_to_be_removed.push(chunk);
 
-                        // NOTE: As usual, these are vectors of `ChunkId`s, as it is legal to
-                        // have perfectly overlapping chunks.
-                        let time_range = time_column.time_range();
-                        per_time
-                            .entry(time_range.min())
-                            .or_default()
-                            .push(chunk.id());
-                        if time_range.min() != time_range.max() {
-                            per_time
-                                .entry(time_range.max())
-                                .or_default()
-                                .push(chunk.id());
-                        }
-                    }
-                }
-
-                // NOTE: There is no point in spending more than a fourth of the time budget on the
-                // mark phase or there is no way the sweep phase will have any time to do anything
-                // with the results anyhow.
-                if start_time.elapsed() >= options.time_budget / 4 || num_bytes_to_drop <= 0.0 {
+                // We divide the time budget equally between the mark and sweep phases.
+                if start_time.elapsed() >= options.time_budget / 2 || num_bytes_to_drop <= 0.0 {
                     break;
                 }
             }
+
+            chunks_to_be_removed
+        };
+
+        {
+            re_tracing::profile_scope!("sweep");
+            self.remove_chunks(
+                chunks_to_be_removed,
+                Some((start_time, options.time_budget)),
+            )
         }
+    }
+
+    /// Surgically removes a set of _temporal_ [`ChunkId`]s from all indices.
+    ///
+    /// This is orders of magnitude faster than trying to `retain()` on all our internal indices,
+    /// when you already know where these chunks live.
+    pub(crate) fn remove_chunks(
+        &mut self,
+        chunks_to_be_removed: Vec<Arc<Chunk>>,
+        time_budget: Option<(Instant, Duration)>,
+    ) -> Vec<ChunkStoreDiff> {
+        re_tracing::profile_function!();
 
         let Self {
             id: _,
             config: _,
-            time_type_registry: _,
-            type_registry: _,
-            per_column_metadata: _, // column metadata is additive only
+            time_type_registry: _,  // purely additive
+            type_registry: _,       // purely additive
+            per_column_metadata: _, // purely additive only
             chunks_per_chunk_id,
             chunk_ids_per_min_row_id,
-            temporal_chunk_ids_per_entity_per_component,
-            temporal_chunk_ids_per_entity,
-            temporal_chunks_stats: _,
+            temporal_chunk_ids_per_entity_per_component: _, // purely additive: virtual index
+            temporal_chunk_ids_per_entity: _,               // purely additive: virtual index
+            temporal_physical_chunks_stats,
             static_chunk_ids_per_entity: _, // we don't GC static data
             static_chunks_stats: _,         // we don't GC static data
             insert_id: _,
@@ -388,357 +397,33 @@ impl ChunkStore {
             event_id: _,
         } = self;
 
-        {
-            re_tracing::profile_scope!("dangling");
-
-            let mut diffs = Vec::new();
-
-            // NOTE: Dangling chunks should never happen: it is the job of the GC to ensure that.
-            //
-            // In release builds, we still want to do the nice thing and clean them up as best as we
-            // can in order to prevent OOMs.
-            //
-            // We should really never be in there, so don't bother accounting that in the time
-            // budget.
-            debug_assert!(
-                chunk_ids_dangling.is_empty(),
-                "detected dangling chunks -- there's a GC bug"
-            );
-            if !chunk_ids_dangling.is_empty() {
-                re_tracing::profile_scope!("dangling");
-
-                chunk_ids_per_min_row_id
-                    .retain(|_row_id, chunk_id| !chunk_ids_dangling.contains(chunk_id));
-
-                // Component-less indices
-                for temporal_chunk_ids_per_timeline in temporal_chunk_ids_per_entity.values_mut() {
-                    for temporal_chunk_ids_per_time in temporal_chunk_ids_per_timeline.values_mut()
-                    {
-                        let ChunkIdSetPerTime {
-                            max_interval_length: _,
-                            per_start_time,
-                            per_end_time,
-                        } = temporal_chunk_ids_per_time;
-
-                        // TODO(cmc): Technically, the optimal thing to do would be to
-                        // recompute `max_interval_length` per time here.
-                        // In practice, this adds a lot of complexity for likely very little
-                        // performance benefit, since we expect the chunks to have similar
-                        // interval lengths on the happy path.
-
-                        for chunk_ids in per_start_time.values_mut() {
-                            chunk_ids.retain(|chunk_id| !chunk_ids_dangling.contains(chunk_id));
-                        }
-                        for chunk_ids in per_end_time.values_mut() {
-                            chunk_ids.retain(|chunk_id| !chunk_ids_dangling.contains(chunk_id));
-                        }
-                    }
-                }
-
-                // Per-component indices
-                for temporal_chunk_ids_per_component in
-                    temporal_chunk_ids_per_entity_per_component.values_mut()
-                {
-                    for temporal_chunk_ids_per_timeline in
-                        temporal_chunk_ids_per_component.values_mut()
-                    {
-                        for temporal_chunk_ids_per_time in
-                            temporal_chunk_ids_per_timeline.values_mut()
-                        {
-                            let ChunkIdSetPerTime {
-                                max_interval_length: _,
-                                per_start_time,
-                                per_end_time,
-                            } = temporal_chunk_ids_per_time;
-
-                            // TODO(cmc): Technically, the optimal thing to do would be to
-                            // recompute `max_interval_length` per time here.
-                            // In practice, this adds a lot of complexity for likely very little
-                            // performance benefit, since we expect the chunks to have similar
-                            // interval lengths on the happy path.
-
-                            for chunk_ids in per_start_time.values_mut() {
-                                chunk_ids.retain(|chunk_id| !chunk_ids_dangling.contains(chunk_id));
-                            }
-                            for chunk_ids in per_end_time.values_mut() {
-                                chunk_ids.retain(|chunk_id| !chunk_ids_dangling.contains(chunk_id));
-                            }
-                        }
-                    }
-                }
-
-                diffs.extend(
-                    chunk_ids_dangling
-                        .into_iter()
-                        .filter_map(|chunk_id| chunks_per_chunk_id.remove(&chunk_id))
-                        .map(ChunkStoreDiff::deletion),
-                );
-            }
-
-            if !chunk_ids_to_be_removed.is_empty() {
-                re_tracing::profile_scope!("sweep");
-
-                diffs.extend(self.remove_chunks(
-                    chunk_ids_to_be_removed,
-                    Some((start_time, options.time_budget)),
-                ));
-            }
-
-            diffs
-        }
-    }
-
-    /// Surgically removes a _temporal_ [`ChunkId`] from all indices.
-    ///
-    /// This is orders of magnitude faster than trying to `retain()` on all our internal indices.
-    ///
-    /// See also [`ChunkStore::remove_chunks`].
-    pub(crate) fn remove_chunk(&mut self, chunk_id: ChunkId) -> Vec<ChunkStoreDiff> {
-        re_tracing::profile_function!();
-
-        let Some(chunk) = self.chunks_per_chunk_id.get(&chunk_id) else {
-            return Vec::new();
-        };
-
-        let mut chunk_ids_to_be_removed =
-            RemovableChunkIdPerTimePerComponentPerTimelinePerEntity::default();
-
-        {
-            let chunk_ids_to_be_removed = chunk_ids_to_be_removed
-                .entry(chunk.entity_path().clone())
-                .or_default();
-
-            for (timeline, time_range_per_component) in chunk.time_range_per_component() {
-                let chunk_ids_to_be_removed = chunk_ids_to_be_removed.entry(timeline).or_default();
-
-                for (component, time_range) in time_range_per_component {
-                    let chunk_ids_to_be_removed =
-                        chunk_ids_to_be_removed.entry(component).or_default();
-
-                    chunk_ids_to_be_removed
-                        .entry(time_range.min())
-                        .or_default()
-                        .push(chunk.id());
-                    chunk_ids_to_be_removed
-                        .entry(time_range.max())
-                        .or_default()
-                        .push(chunk.id());
-                }
-            }
-        }
-
-        self.remove_chunks(chunk_ids_to_be_removed, None)
-    }
-
-    /// Surgically removes a set of _temporal_ [`ChunkId`]s from all indices.
-    ///
-    /// This is orders of magnitude faster than trying to `retain()` on all our internal indices,
-    /// when you already know where these chunks live.
-    ///
-    /// See also [`ChunkStore::remove_chunk`].
-    pub(crate) fn remove_chunks(
-        &mut self,
-        chunk_ids_to_be_removed: RemovableChunkIdPerTimePerComponentPerTimelinePerEntity,
-        time_budget: Option<(Instant, Duration)>,
-    ) -> Vec<ChunkStoreDiff> {
-        re_tracing::profile_function!();
-
-        // NOTE: We cannot blindly `retain` across all temporal tables, it's way too costly
-        // and slow. Rather we need to surgically remove the superfluous chunks.
-
-        let mut chunk_ids_removed = HashSet::default();
-
-        // Because we have both a per-component and a component-less index that refer to the same
-        // chunks, we must make sure that they get garbage collected in sync.
-        // That implies making sure that we don't run out of time budget after we've GC'd one but
-        // before we had time to clean the other.
-
-        for (entity_path, chunk_ids_to_be_removed) in chunk_ids_to_be_removed {
-            re_tracing::profile_scope!("chunk-id");
-
-            let HashMapEntry::Occupied(mut temporal_chunk_ids_per_timeline) = self
-                .temporal_chunk_ids_per_entity_per_component
-                .entry(entity_path.clone())
-            else {
-                continue;
-            };
-
-            let HashMapEntry::Occupied(mut temporal_chunk_ids_per_timeline_componentless) =
-                self.temporal_chunk_ids_per_entity.entry(entity_path)
-            else {
-                continue;
-            };
-
-            for (timeline, chunk_ids_to_be_removed) in chunk_ids_to_be_removed {
-                re_tracing::profile_scope!("timeline");
-                // Component-less indices
-                {
-                    let HashMapEntry::Occupied(mut temporal_chunk_ids_per_time_componentless) =
-                        temporal_chunk_ids_per_timeline_componentless
-                            .get_mut()
-                            .entry(timeline)
-                    else {
-                        continue;
-                    };
-
-                    let ChunkIdSetPerTime {
-                        max_interval_length: _,
-                        per_start_time,
-                        per_end_time,
-                    } = temporal_chunk_ids_per_time_componentless.get_mut();
-
-                    // TODO(cmc): Technically, the optimal thing to do would be to
-                    // recompute `max_interval_length` per time here.
-                    // In practice, this adds a lot of complexity for likely very little
-                    // performance benefit, since we expect the chunks to have similar
-                    // interval lengths on the happy path.
-
-                    for chunk_ids_to_be_removed in chunk_ids_to_be_removed.values() {
-                        for (&time, chunk_ids) in chunk_ids_to_be_removed {
-                            if let BTreeMapEntry::Occupied(mut chunk_id_set) =
-                                per_start_time.entry(time)
-                            {
-                                for chunk_id in chunk_ids {
-                                    chunk_id_set.get_mut().remove(chunk_id);
-                                }
-                                if chunk_id_set.get().is_empty() {
-                                    chunk_id_set.remove_entry();
-                                }
-                            }
-
-                            if let BTreeMapEntry::Occupied(mut chunk_id_set) =
-                                per_end_time.entry(time)
-                            {
-                                for chunk_id in chunk_ids {
-                                    chunk_id_set.get_mut().remove(chunk_id);
-                                }
-                                if chunk_id_set.get().is_empty() {
-                                    chunk_id_set.remove_entry();
-                                }
-                            }
-
-                            chunk_ids_removed.extend(chunk_ids);
-                        }
-
-                        if let Some((start_time, time_budget)) = time_budget
-                            && start_time.elapsed() >= time_budget
-                        {
-                            break;
-                        }
-                    }
-
-                    if per_start_time.is_empty() && per_end_time.is_empty() {
-                        temporal_chunk_ids_per_time_componentless.remove_entry();
-                    }
-                }
-
-                // Per-component indices
-                //
-                // NOTE: This must go all the way, no matter the time budget left. Otherwise the
-                // component-less and per-component indices would go out of sync.
-
-                let HashMapEntry::Occupied(mut temporal_chunk_ids_per_component) =
-                    temporal_chunk_ids_per_timeline.get_mut().entry(timeline)
-                else {
-                    continue;
-                };
-
-                for (component_descr, chunk_ids_to_be_removed) in chunk_ids_to_be_removed {
-                    let HashMapEntry::Occupied(mut temporal_chunk_ids_per_time) =
-                        temporal_chunk_ids_per_component
-                            .get_mut()
-                            .entry(component_descr)
-                    else {
-                        continue;
-                    };
-
-                    let ChunkIdSetPerTime {
-                        max_interval_length: _,
-                        per_start_time,
-                        per_end_time,
-                    } = temporal_chunk_ids_per_time.get_mut();
-
-                    // TODO(cmc): Technically, the optimal thing to do would be to
-                    // recompute `max_interval_length` per time here.
-                    // In practice, this adds a lot of complexity for likely very little
-                    // performance benefit, since we expect the chunks to have similar
-                    // interval lengths on the happy path.
-
-                    for (time, chunk_ids) in chunk_ids_to_be_removed {
-                        if let BTreeMapEntry::Occupied(mut chunk_id_set) =
-                            per_start_time.entry(time)
-                        {
-                            for chunk_id in chunk_ids
-                                .iter()
-                                .filter(|chunk_id| chunk_ids_removed.contains(*chunk_id))
-                            {
-                                chunk_id_set.get_mut().remove(chunk_id);
-                            }
-                            if chunk_id_set.get().is_empty() {
-                                chunk_id_set.remove_entry();
-                            }
-                        }
-
-                        if let BTreeMapEntry::Occupied(mut chunk_id_set) = per_end_time.entry(time)
-                        {
-                            for chunk_id in chunk_ids
-                                .iter()
-                                .filter(|chunk_id| chunk_ids_removed.contains(*chunk_id))
-                            {
-                                chunk_id_set.get_mut().remove(chunk_id);
-                            }
-                            if chunk_id_set.get().is_empty() {
-                                chunk_id_set.remove_entry();
-                            }
-                        }
-                    }
-
-                    if per_start_time.is_empty() && per_end_time.is_empty() {
-                        temporal_chunk_ids_per_time.remove_entry();
-                    }
-                }
-
-                if temporal_chunk_ids_per_component.get().is_empty() {
-                    temporal_chunk_ids_per_component.remove_entry();
-                }
-            }
-
-            if temporal_chunk_ids_per_timeline.get().is_empty() {
-                temporal_chunk_ids_per_timeline.remove_entry();
-            }
-
-            if temporal_chunk_ids_per_timeline_componentless
-                .get()
-                .is_empty()
+        let mut diffs = Vec::with_capacity(chunks_to_be_removed.len());
+        for chunk in chunks_to_be_removed {
+            if let Some((start_time, time_budget)) = time_budget
+                && start_time.elapsed() >= time_budget
             {
-                temporal_chunk_ids_per_timeline_componentless.remove_entry();
+                break;
             }
+
+            if let Some(row_id_min) = chunk.row_id_range().map(|(min, _)| min) {
+                chunk_ids_per_min_row_id.remove(&row_id_min);
+            }
+            let Some(chunk) = chunks_per_chunk_id.remove(&chunk.id()) else {
+                continue;
+            };
+
+            // TODO(cmc): Technically, the optimal thing to do would be to recompute
+            // `max_interval_length` per time here.
+            // In practice, this adds a lot of complexity for likely very little
+            // performance benefit, since we expect the chunks to have similar interval
+            // lengths on the happy path.
+
+            *temporal_physical_chunks_stats -= ChunkStoreChunkStats::from_chunk(&chunk);
+
+            diffs.push(ChunkStoreDiff::deletion(chunk));
         }
 
-        let min_row_ids_removed = chunk_ids_removed.iter().filter_map(|chunk_id| {
-            let chunk = self.chunks_per_chunk_id.get(chunk_id)?;
-            chunk.row_id_range().map(|(min, _)| min)
-        });
-        for row_id in min_row_ids_removed {
-            if self.chunk_ids_per_min_row_id.remove(&row_id).is_none() {
-                re_log::warn!(
-                    %row_id,
-                    "Row ID marked for removal was not found, there's bug in the Chunk Store"
-                );
-            }
-        }
-
-        {
-            re_tracing::profile_scope!("last collect");
-            chunk_ids_removed
-                .into_iter()
-                .filter_map(|chunk_id| self.chunks_per_chunk_id.remove(&chunk_id))
-                .inspect(|chunk| {
-                    self.temporal_chunks_stats -= ChunkStoreChunkStats::from_chunk(chunk);
-                })
-                .map(ChunkStoreDiff::deletion)
-                .collect()
-        }
+        diffs
     }
 }
 
@@ -788,10 +473,14 @@ mod tests {
             let mut store = setup_store();
 
             assert_eq!(NUM_CHUNKS as usize, store.num_chunks());
-            store.gc(&GarbageCollectionOptions {
-                furthest_from: Some((TimelineName::log_tick(), TimeInt::new_temporal(pivot))),
-                ..GarbageCollectionOptions::gc_everything()
-            });
+            for _ in 0..3 {
+                // Call `store.gc()` more than once just to make sure nothing weird happens with
+                // all the shadow indices left by the first call.
+                store.gc(&GarbageCollectionOptions {
+                    furthest_from: Some((TimelineName::log_tick(), TimeInt::new_temporal(pivot))),
+                    ..GarbageCollectionOptions::gc_everything()
+                });
+            }
             assert_eq!(0, store.num_chunks());
         }
     }
