@@ -666,17 +666,31 @@ impl EntityDb {
 
         let store_events = self.gc(&GarbageCollectionOptions {
             target: GarbageCollectionTarget::DropAtLeastFraction(fraction_to_purge as _),
-            protect_latest: 1,
             time_budget: DEFAULT_GC_TIME_BUDGET,
+
+            // NOTE: This will only apply if the GC is forced to fall back to row ID based collection,
+            // otherwise timestamp-based collection will ignore it.
+            protect_latest: 1,
 
             // TODO(emilk): we could protect the data that is currently being viewed
             // (e.g. when paused in the live camera example).
             // To be perfect it would need margins (because of latest-at), i.e. we would need to know
             // exactly how far back the latest-at is of each component at the current time…
             // …but maybe it doesn't have to be perfect.
+            //
+            // NOTE: This will only apply if the GC is forced to fall back to row ID based collection,
+            // otherwise timestamp-based collection will ignore it.
             protected_time_ranges: Default::default(),
 
-            furthest_from: time_cursor.map(|(timeline, time)| (*timeline.name(), time)),
+            furthest_from: if self.rrd_manifest_index.has_manifest() {
+                // If we have an RRD manifest, it means we can download chunks on-demand.
+                // So it makes sense to GC the things furthest from the current time cursor:
+                time_cursor.map(|(timeline, time)| (*timeline.name(), time))
+            } else {
+                // If we don't have an RRD manifest, then we can't redownload data,
+                // and we GC the oldest data instead.
+                None
+            },
         });
 
         if store_events.is_empty() {
