@@ -2297,8 +2297,10 @@ pub mod rerun_cloud_service_client {
         pub async fn get_rrd_manifest(
             &mut self,
             request: impl tonic::IntoRequest<super::GetRrdManifestRequest>,
-        ) -> std::result::Result<tonic::Response<super::GetRrdManifestResponse>, tonic::Status>
-        {
+        ) -> std::result::Result<
+            tonic::Response<tonic::codec::Streaming<super::GetRrdManifestResponse>>,
+            tonic::Status,
+        > {
             self.inner.ready().await.map_err(|e| {
                 tonic::Status::unknown(format!("Service was not ready: {}", e.into()))
             })?;
@@ -2311,7 +2313,7 @@ pub mod rerun_cloud_service_client {
                 "rerun.cloud.v1alpha1.RerunCloudService",
                 "GetRrdManifest",
             ));
-            self.inner.unary(req, path, codec).await
+            self.inner.server_streaming(req, path, codec).await
         }
         /// Creates a custom index for a specific column (vector search, full-text search, etc).
         ///
@@ -2769,13 +2771,18 @@ pub mod rerun_cloud_service_server {
             &self,
             request: tonic::Request<super::GetDatasetSchemaRequest>,
         ) -> std::result::Result<tonic::Response<super::GetDatasetSchemaResponse>, tonic::Status>;
+        /// Server streaming response type for the GetRrdManifest method.
+        type GetRrdManifestStream: tonic::codegen::tokio_stream::Stream<
+                Item = std::result::Result<super::GetRrdManifestResponse, tonic::Status>,
+            > + std::marker::Send
+            + 'static;
         /// Get the RRD Footer manifest.
         ///
         /// This includes details about what chunks there are, and what kind of data they contain.
         async fn get_rrd_manifest(
             &self,
             request: tonic::Request<super::GetRrdManifestRequest>,
-        ) -> std::result::Result<tonic::Response<super::GetRrdManifestResponse>, tonic::Status>;
+        ) -> std::result::Result<tonic::Response<Self::GetRrdManifestStream>, tonic::Status>;
         /// Creates a custom index for a specific column (vector search, full-text search, etc).
         ///
         /// This endpoint requires the standard dataset headers.
@@ -3675,11 +3682,13 @@ pub mod rerun_cloud_service_server {
                     #[allow(non_camel_case_types)]
                     struct GetRrdManifestSvc<T: RerunCloudService>(pub Arc<T>);
                     impl<T: RerunCloudService>
-                        tonic::server::UnaryService<super::GetRrdManifestRequest>
+                        tonic::server::ServerStreamingService<super::GetRrdManifestRequest>
                         for GetRrdManifestSvc<T>
                     {
                         type Response = super::GetRrdManifestResponse;
-                        type Future = BoxFuture<tonic::Response<Self::Response>, tonic::Status>;
+                        type ResponseStream = T::GetRrdManifestStream;
+                        type Future =
+                            BoxFuture<tonic::Response<Self::ResponseStream>, tonic::Status>;
                         fn call(
                             &mut self,
                             request: tonic::Request<super::GetRrdManifestRequest>,
@@ -3708,7 +3717,7 @@ pub mod rerun_cloud_service_server {
                                 max_decoding_message_size,
                                 max_encoding_message_size,
                             );
-                        let res = grpc.unary(method, req).await;
+                        let res = grpc.server_streaming(method, req).await;
                         Ok(res)
                     };
                     Box::pin(fut)
