@@ -182,7 +182,10 @@ fn setup_blueprint(
         ctx.save_blueprint_archetype(
             cos_override_path.clone(),
             // Override which visualizer to use for the `cos` plot.
-            &blueprint::archetypes::VisualizerOverrides::new(["SeriesPoints"]),
+            &blueprint::archetypes::VisualizerOverrides::new([
+                // TODO(RR-3153): remove the `as_str()`.
+                archetypes::SeriesPoints::visualizer().as_str(),
+            ]),
         );
         ctx.save_blueprint_archetype(
             cos_override_path,
