@@ -669,6 +669,7 @@ impl App {
     }
 
     fn run_pending_system_commands(&mut self, store_hub: &mut StoreHub, egui_ctx: &egui::Context) {
+        re_tracing::profile_function!();
         while let Some((from_where, cmd)) = self.command_receiver.recv_system() {
             self.run_system_command(from_where, cmd, store_hub, egui_ctx);
         }
@@ -813,6 +814,8 @@ impl App {
         store_hub: &mut StoreHub,
         egui_ctx: &egui::Context,
     ) {
+        re_tracing::profile_function!(cmd.debug_name());
+
         match cmd {
             SystemCommand::TimeControlCommands {
                 store_id,
@@ -1302,6 +1305,8 @@ impl App {
         egui_ctx: &egui::Context,
         data_source: &LogDataSource,
     ) {
+        re_tracing::profile_function!();
+
         // Check if we've already loaded this data source and should just switch to it.
         //
         // Go through all sources that are still loading and those that are already in the store_hub.
@@ -2334,7 +2339,9 @@ impl App {
                 }
             });
 
-        self.notifications.show_toasts(egui_ctx);
+        if self.app_options().show_notification_toasts {
+            self.notifications.show_toasts(egui_ctx);
+        }
     }
 
     /// Show recent text log messages to the user as toast notifications.
@@ -3233,20 +3240,19 @@ impl eframe::App for App {
             force_store_info,
             promise,
         }) = &self.open_files_promise
+            && let Some(files) = promise.ready()
         {
-            if let Some(files) = promise.ready() {
-                for file in files {
-                    self.command_sender
-                        .send_system(SystemCommand::LoadDataSource(LogDataSource::FileContents(
-                            FileSource::FileDialog {
-                                recommended_store_id: recommended_store_id.clone(),
-                                force_store_info: *force_store_info,
-                            },
-                            file.clone(),
-                        )));
-                }
-                self.open_files_promise = None;
+            for file in files {
+                self.command_sender
+                    .send_system(SystemCommand::LoadDataSource(LogDataSource::FileContents(
+                        FileSource::FileDialog {
+                            recommended_store_id: recommended_store_id.clone(),
+                            force_store_info: *force_store_info,
+                        },
+                        file.clone(),
+                    )));
             }
+            self.open_files_promise = None;
         }
 
         // NOTE: GPU resource stats are cheap to compute so we always do.
@@ -3702,7 +3708,7 @@ fn save_blueprint(app: &mut App, store_context: Option<&StoreContext<'_>>) -> an
 
 // TODO(emilk): unify this with `ViewerContext::save_file_dialog`
 #[allow(clippy::allow_attributes, clippy::needless_pass_by_ref_mut)] // `app` is only used on native
-#[allow(clippy::allow_attributes, clippy::unnecessary_wraps)] // cannot return error on web
+#[allow(clippy::unnecessary_wraps)] // cannot return error on web
 fn save_entity_db(
     #[allow(clippy::allow_attributes, unused_variables)] app: &mut App, // only used on native
     rrd_version: CrateVersion,
